Auto update on start
We can skip trying to create a "check for updates" button within the program. By forcing the updater script to run as part of the startup script, then we get automatic updates via CTRL+Q, which kills the program, therefore restarting the getty session, therefore re-running the script and checking for and installing updates! The updater script then restarts getty anyways, so the new version gets run without ever needing a reboot. How fucking cheeky and clever is that?