Remove mp3 version of file
This may be important in the future. It seems that I can store the mp3 in the Github repository, but not the wav.
So maybe I should consider reading an mp3 rather than a wav, keep the wav for the commercial system with proper speakers for better audio quality.
Ooh, yes! A new commercial feature!